working tic tac toe game
this is huge. created arrays for each player leds, player led states, used timer to check for win or tie every second, removed switch case, fixed SETP6 and SETP4 macros, created if statement to reset the game, implemented two for loops to verify the state of the leds (sometimes they would get stuck on or off due to button interrupts).